requirements.txt
accelerate==0.24.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
arxiv==2.1.0
async-timeout==4.0.3
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
blis==0.7.11
brotli==1.2.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
cn2an==0.5.23
coloredlogs==15.0.1
confection==0.1.5
cryptography==46.0.6
cymem==2.0.13
dashscope==1.13.6
dataclasses==0.6
dataclasses-json==0.6.7
Deprecated==1.3.1
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
duckduckgo-search==3.9.11
effdet==0.4.1
einops==0.8.2
emoji==2.15.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss==1.10.0
fastapi==0.108.0
feedparser==6.0.10
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fschat==0.2.34
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.25.2
httpx-sse==0.4.3
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==6.11.0
iniconfig==2.3.0
iopath==0.1.10
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jq==1.6.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.0.354
langchain-classic==1.0.3
langchain-community==0.0.20
langchain-core==0.1.53
langchain-deepseek==0.1.4
langchain-experimental==0.0.47
langchain-openai==0.1.0
langchain-text-splitters==1.1.1
langcodes==3.5.1
langdetect==1.0.9
langsmith==0.1.147
latex2mathml==3.79.0
layoutparser==0.3.4
llama-index==0.10.68
llama-index-agent-openai==0.2.1
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llama_cloud==1.6.0
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
markdown2==2.5.5
markdownify==0.11.6
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
metaphor-python==0.1.23
ml_dtypes==0.5.4
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
mysql-connector-python==8.3.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.3.4
nltk==3.9.4
numexpr==2.8.8
numpy==1.24.4
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.15.1
openai==1.109.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.8
packaging==23.2
pageindex==0.2.8
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
Pillow==9.5.0
pluggy==1.6.0
portalocker==3.2.0
preshed==3.0.13
proces==0.1.7
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.9
psutil==5.9.8
pyarrow==23.0.1
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==3.0
pydantic==1.10.13
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.12.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pypandoc==1.17
pypdf==4.3.1
pypdfium2==5.6.0
pypinyin==0.50.0
pysrt==1.1.2
pytesseract==0.3.13
pytest==7.4.4
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-pptx==0.6.23
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
rapidocr-onnxruntime==1.3.8
redis==7.4.0
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
safetensors==0.4.5
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.1
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.20.2
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.19
srsly==2.5.3
sse-starlette==1.8.2
starlette==0.32.0.post1
streamlit==1.29.0
streamlit-aggrid==1.2.1.post2
streamlit-antd-components==0.3.0
streamlit-chatbox==1.1.11
streamlit-drawable-canvas==0.9.3
streamlit-feedback==0.1.4
streamlit-markdown==1.1.0
streamlit-modal==0.1.2
streamlit-option-menu==0.3.6
striprtf==0.0.26
strsimpy==0.2.1
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.10.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.5.2
timm==1.0.26
tokenizers==0.15.2
toml==0.10.2
tomli==2.4.1
toolz==1.1.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.5.5
tqdm==4.67.3
transformers==4.36.2
transformers-stream-generator==0.0.4
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unstructured==0.11.0
unstructured-inference==0.7.15
unstructured.pytesseract==0.3.15
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
wavedrom==2.0.3.post3
wcwidth==0.6.0
weasel==0.4.3
websockets==16.0
wrapt==2.1.2
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
youtube-search==2.1.2
zhipuai==1.0.7
zipp==3.23.0
zstandard==0.25.0
